// src/api/model/notificationResponseType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */

export type NotificationResponseType = typeof NotificationResponseType[keyof typeof NotificationResponseType];


export const NotificationResponseType = {
  TRACE_REVEALED: 'TRACE_REVEALED',
  FRIEND_REQUEST_ACCEPT: 'FRIEND_REQUEST_ACCEPT',
  NOTICE: 'NOTICE',
  BUZZ_ARRIVAL: 'BUZZ_ARRIVAL',
  BUZZ_RESPONSE: 'BUZZ_RESPONSE',
  FLAG_ENCORE: 'FLAG_ENCORE',
  FLAG_CANCELED: 'FLAG_CANCELED',
  FLAG_SCHEDULE_CHANGED: 'FLAG_SCHEDULE_CHANGED',
} as const;